requirements.txt
absl-py==0.9.0
actionlib==1.11.13
angles==1.9.12
base-local-planner==1.14.7
bondpy==1.8.3
cachetools==4.1.1
camera-calibration==1.12.23
camera-calibration-parsers==1.11.13
capabilities==0.2.0
catkin==0.7.20
catkin-pkg==0.4.22
certifi==2020.6.20
chardet==3.0.4
controller-manager==0.13.5
controller-manager-msgs==0.13.5
controller-manager-tests==0.13.5
create-dashboard==2.3.1
create-driver==2.3.1
create-node==2.3.1
cv-bridge==1.12.8
diagnostic-analysis==1.9.3
diagnostic-common-diagnostics==1.9.3
diagnostic-updater==1.9.3
distro==1.5.0
docutils==0.16
dynamic-reconfigure==1.5.50
gazebo-plugins==2.5.20
gazebo-ros==2.5.20
gencpp==0.6.0
geneus==2.2.6
genlisp==0.4.16
genmsg==0.5.11
gennodejs==2.0.1
genpy==0.6.7
google-auth==1.19.2
google-auth-oauthlib==0.4.1
grpcio==1.30.0
idna==2.10
image-geometry==1.12.8
importlib-metadata==1.7.0
interactive-markers==1.11.5
joint-state-publisher==1.12.15
kobuki-dashboard==0.5.7
laser-geometry==1.6.5
Markdown==3.2.2
message-filters==1.12.14
numpy==1.18.5
oauthlib==3.1.0
openni2-launch==0.4.2
Pillow==7.2.0
pluginlib==1.11.3
protobuf==3.12.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.7
python-dateutil==2.8.1
python-qt-binding==0.3.7
PyYAML==5.3.1
qt-dotgraph==0.3.17
qt-gui==0.3.17
qt-gui-cpp==0.3.17
qt-gui-py-common==0.3.17
requests==2.24.0
requests-oauthlib==1.3.0
resource-retriever==1.12.6
rocon-app-manager==0.9.1
rocon-app-utilities==0.9.1
rocon-apps==0.9.1
rocon-console==0.3.2
rocon-ebnf==0.3.2
rocon-gateway==0.8.1
rocon-gateway-utils==0.8.1
rocon-hub==0.8.1
rocon-hub-client==0.8.1
rocon-interactions==0.3.2
rocon-master-info==0.3.2
rocon-python-comms==0.3.2
rocon-python-redis==0.3.2
rocon-python-utils==0.3.2
rocon-python-wifi==0.3.2
rocon-semantic-version==0.3.2
rocon-uri==0.3.2
rosbag==1.12.14
rosboost-cfg==1.14.6
rosclean==1.14.6
roscreate==1.14.6
rosgraph==1.12.14
roslaunch==1.12.14
roslib==1.14.6
roslint==0.11.0
roslz4==1.12.14
rosmake==1.14.6
rosmaster==1.12.14
rosmsg==1.12.14
rosnode==1.12.14
rosparam==1.12.14
rospkg==1.2.8
rospy==1.12.14
rospy-message-converter==0.5.0
rosservice==1.12.14
rostest==1.12.14
rostopic==1.12.14
rosunit==1.14.6
roswtf==1.12.14
rqt-action==0.4.9
rqt-bag==0.4.12
rqt-bag-plugins==0.4.12
rqt-console==0.4.9
rqt-dep==0.4.9
rqt-graph==0.4.11
rqt-gui==0.5.0
rqt-gui-py==0.5.0
rqt-image-view==0.4.14
rqt-joint-trajectory-controller==0.13.6
rqt-launch==0.4.8
rqt-logger-level==0.4.8
rqt-moveit==0.5.7
rqt-msg==0.4.8
rqt-nav-view==0.5.7
rqt-plot==0.4.8
rqt-pose-view==0.5.8
rqt-publisher==0.4.8
rqt-py-common==0.5.0
rqt-py-console==0.4.8
rqt-reconfigure==0.5.1
rqt-robot-dashboard==0.5.7
rqt-robot-monitor==0.5.8
rqt-robot-steering==0.5.9
rqt-runtime-monitor==0.5.7
rqt-rviz==0.5.10
rqt-service-caller==0.4.8
rqt-shell==0.4.9
rqt-srv==0.4.8
rqt-tf-tree==0.6.0
rqt-top==0.4.8
rqt-topic==0.4.11
rqt-web==0.4.8
rsa==4.6
rviz==1.12.17
sensor-msgs==1.12.7
Shapely==1.7.0
six==1.15.0
smach==2.0.1
smach-ros==2.0.1
smclib==1.8.3
tensorboard==2.1.0
tensorboard-plugin-wit==1.7.0
tf==1.11.9
tf-conversions==1.11.9
tf2-geometry-msgs==0.5.20
tf2-kdl==0.5.20
tf2-py==0.5.20
tf2-ros==0.5.20
topic-tools==1.12.14
torch==1.2.0
torchvision==0.4.0
unique-id==1.0.5
urllib3==1.25.10
Werkzeug==1.0.1
world-canvas-server==0.2.0
world-canvas-utils==0.2.0
xacro==1.11.3
zipp==1.2.0
